CBC-994 Improve code with clippy

+++ crates/ethcore/src/engines/clique/params.rs
@@ -28,8 +28,8 @@ pub struct CliqueParams {
 
 impl From<ethjson::spec::CliqueParams> for CliqueParams {
     fn from(p: ethjson::spec::CliqueParams) -> Self {
-        let period = p.period.map_or_else(|| 30000 as u64, Into::into);
-        let epoch = p.epoch.map_or_else(|| 15 as u64, Into::into);
+        let period = p.period.map_or_else(|| 30000_u64, Into::into);
+        let epoch = p.epoch.map_or_else(|| 15_u64, Into::into);
 
         assert!(epoch > 0);
